Move Task model to subdirectory and add casts
